% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportd_darkblue.R
\name{reportd_darkblue}
\alias{reportd_darkblue}
\title{Function for green and blue dark theme for GSS reports in the form of a RMarkdown file.}
\usage{
reportd_darkblue(...)
}
\arguments{
\item{...}{Arguments passed on to pagedown::html_paged}
}
\value{
Pagedown dark theme template that was designed for creating reports for the Graduate Student Society at UBC.
}
\description{
Function for green and blue dark theme for GSS reports in the form of a RMarkdown file.
}
