% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportd_darkpink.R
\name{reportd_darkpink}
\alias{reportd_darkpink}
\title{Creates a dark theme, green and pink as main colors in the form of a RMarkdown file}
\usage{
reportd_darkpink(...)
}
\arguments{
\item{...}{Arguments passed on to pagedown::html_paged}
}
\value{
Pagedown template that was designed for creating reports for the Graduate Student Society at UBC.
}
\description{
Creates a dark theme, green and pink as main colors in the form of a RMarkdown file
}
