% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportd_light.R
\name{reportd_light}
\alias{reportd_light}
\title{Function for the light theme for GSS reports}
\usage{
reportd_light(...)
}
\arguments{
\item{...}{}
}
\value{
a pagedown template
}
\description{
Function for the light theme for GSS reports
}
