% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportd_light_v2.R
\name{reportd_light_v2}
\alias{reportd_light_v2}
\title{Creates another light theme for GSS reports in the form a a RMarkdown file}
\usage{
reportd_light_v2(...)
}
\arguments{
\item{...}{Arguments passed on to pagedown::html_paged}
}
\value{
Pagedown template that was designed for creating reports for the Graduate Student Society at UBC.
}
\description{
Creates another light theme for GSS reports in the form a a RMarkdown file
}
